requirements.txt
PSNAWP~=2.0.0
pypresence~=4.2.0
PyYAML~=5.4.1